Fix test dependencies to work with clean venv
For some reason, it doesn't like when I specify a newer setuptools, and
to use newer pytest-runner, I need a MUCH newer setuptools version. So
lets roll it all back to something that plays nice, at least for now,
and perhaps only for Ubuntu 16.04. But I can deal with that later.